pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fenicsx-cosim"
version = "0.1.0"
description = "A native partitioned multiphysics coupling library for FEniCSx"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Lavakumar"},
]
keywords = ["FEniCSx", "co-simulation", "multiphysics", "coupling", "partitioned"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Physics",
]
dependencies = [
    "numpy>=1.24.0",
    "pyzmq>=25.0.0",
    "scipy>=1.10.0",
]

[project.optional-dependencies]
fenicsx = [
    "fenics-dolfinx>=0.10.0",
    "mpi4py",
]
dev = [
    "pytest>=7.0",
    "pytest-mpi",
]

[tool.setuptools.packages.find]
where = ["src"]
